% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_MH.r
\name{autocorr_IMH}
\alias{autocorr_IMH}
\title{Autocorrelation function of the Idependent Metropolis-Hastings with exponential target}
\usage{
autocorr_IMH(k, theta)
}
\arguments{
\item{k}{lag}

\item{theta}{proposal parameter}
}
\value{
the autocorrelation at lag k
}
\description{
Autocorrelation function of the Idependent Metropolis-Hastings with exponential target
}
