% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_LNAR1.r
\name{autocov_LNAR1}
\alias{autocov_LNAR1}
\title{Autocovariance function of LNAR1 process}
\usage{
autocov_LNAR1(k, lmean, lsd, phi)
}
\arguments{
\item{k}{the lag}

\item{lmean}{log mean}

\item{lsd}{log standard deviation}

\item{phi}{correlation parameter of the latent AR(1)}
}
\value{
the autocovariance at \code{k}
}
\description{
Autocovariance function of LNAR1 process
}
