% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_LAR1.r
\name{compute_LAR_alpha}
\alias{compute_LAR_alpha}
\title{Find the 'alpha' parameter of LAR(1) process}
\usage{
compute_LAR_alpha(sigma_e, phi, p)
}
\arguments{
\item{sigma_e}{the standard deviation of the innovations of the latent AR(1)}

\item{phi}{correlation parameter of the latent AR(1)}

\item{p}{the marginal probability that X = 1}
}
\value{
a scalar between 0 and 1
}
\description{
Find the 'alpha' parameter of LAR(1) process
}
\details{
This is the approximation of a LAR(1) by a first-order two-state Markov chain
}
