% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_LNAR1.r
\name{compute_p_LT_variance}
\alias{compute_p_LT_variance}
\title{Compute the long-term (LT) variance of a quantile indicator}
\usage{
compute_p_LT_variance(p, lmean, lsd, phi, K = 5000)
}
\arguments{
\item{p}{a quantile (between 0 and 1)}

\item{lmean}{log mean}

\item{lsd}{log standard deviation}

\item{phi}{correlation parameter of the latent AR(1)}

\item{K}{an upper bound up to which to sum the autocorrelation function.
Default is K = 5000}
}
\value{
a scalar containing the LT variance of the target quantile
}
\description{
Compute the long-term (LT) variance of a quantile indicator
}
