% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_LNAR_pars.r
\name{est_LNAR_pars}
\alias{est_LNAR_pars}
\title{Estimate the parameters of a LNAR process}
\usage{
est_LNAR_pars(X, level = 0.95)
}
\arguments{
\item{X}{draws from the observed process}

\item{level}{confidence level for parameter confidence intervals}
}
\value{
a data.frame with point estimates and confidence intervals for all
three parameters
}
\description{
Estimate the parameters of a LNAR process
}
