% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_utilities.r
\name{exponential_bci}
\alias{exponential_bci}
\title{Compute the central BCI of an exponential distribution}
\usage{
exponential_bci(alpha = 0.95, theta)
}
\arguments{
\item{alpha}{the BCI level  (default is 0.95)}

\item{theta}{the rate parameter.}
}
\value{
a vector of size two with the HPD endpoints
}
\description{
Compute the central BCI of an exponential distribution
}
\examples{
exponential_bci(0.95, 1)
exponential_bci(0.90, 1)
exponential_bci(0.95, 10)
exponential_bci(0.90, 10)
}
