% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_utilities.r
\name{exponential_hpd}
\alias{exponential_hpd}
\title{Compute the HPD of an exponential distribution}
\usage{
exponential_hpd(alpha = 0.95, theta)
}
\arguments{
\item{alpha}{the HPD level  (default is 0.95)}

\item{theta}{the rate parameter.}
}
\value{
a vector of size two with the HPD endpoints
}
\description{
Compute the HPD of an exponential distribution
}
\examples{
exponential_hpd(0.95, 1)
exponential_hpd(0.90, 1)
exponential_hpd(0.95, 10)
exponential_hpd(0.90, 10)
}
