% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_LNAR1.r
\name{find_phi_LNAR1}
\alias{find_phi_LNAR1}
\title{Find the autocorrelation parameter for a given efficiency}
\usage{
find_phi_LNAR1(eff, lmean, lsd)
}
\arguments{
\item{eff}{the desired efficiency}

\item{lmean}{log mean}

\item{lsd}{log standard deviation}
}
\value{
the value of phi (between -1 and 1)
}
\description{
Find the autocorrelation parameter for a given efficiency
}
