% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_LNAR1.r
\name{fit_LNAR1}
\alias{fit_LNAR1}
\title{Fits a log-normal autorregressive (LNAR) model of order 1}
\usage{
fit_LNAR1(X)
}
\arguments{
\item{X}{draws from MCMC}
}
\value{
a list containing the estimated LNAR1 parameters (lmean, lsd and phi)
}
\description{
Fits a log-normal autorregressive (LNAR) model of order 1
}
