% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_MCMC.r
\name{generate_fake_MCMC}
\alias{generate_fake_MCMC}
\title{Generate 'fake' MCMC draws from a latent gaussian autorregressive order 1 (LNAR1) process}
\usage{
generate_fake_MCMC(N = 10000, lmean = 0, lsd = 1, phi = -0.75, LN = TRUE)
}
\arguments{
\item{N}{number of draws}

\item{lmean}{log mean}

\item{lsd}{log standard deviation}

\item{phi}{correlation parameter of the latent AR(1)}

\item{LN}{whether to generate from a log-normal (default is true)}
}
\value{
a vector of size N with the draws
}
\description{
Generate 'fake' MCMC draws from a latent gaussian autorregressive order 1 (LNAR1) process
}
