% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_LAR1.r
\name{generate_LAR_TS}
\alias{generate_LAR_TS}
\title{Sample from a LAR(1) process}
\usage{
generate_LAR_TS(N, phi, target.p)
}
\arguments{
\item{N}{the number of draws}

\item{phi}{correlation parameter of the latent AR(1)}

\item{target.p}{the marginal probability that X = 1}
}
\value{
a vector with \code{N} draws
}
\description{
Sample from a LAR(1) process
}
\examples{
p <- .23
thePhi <- get_LAR_phi(Eff = .6, p = p)
X <- generate_LAR_TS(N = 1E3, phi = thePhi, target.p = p)
}
