% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_TSMC.r
\name{generate_MC_TS}
\alias{generate_MC_TS}
\title{Generate draws from a two state Markov chain}
\usage{
generate_MC_TS(N, target.ess, target.p = 0.5, x0 = NULL)
}
\arguments{
\item{N}{the number of draws}

\item{target.ess}{target effective sample size (ESS)}

\item{target.p}{the marginal probability that X = 1}

\item{x0}{initial state. Either 0, 1 or \code{NULL}.
if \code{NULL}, a fair coin is flipped to pick \code{x0}}
}
\value{
a vector with \code{N}
}
\description{
Generate draws from a two state Markov chain
}
\examples{
X <- generate_MC_TS(N = 1e3, target.ess = 150, target.p = .23)
}
