% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_MH.r
\name{IMH_eff}
\alias{IMH_eff}
\title{Efficiency (integrated autorrelation time) of the IMH with exponential target}
\usage{
IMH_eff(theta)
}
\arguments{
\item{theta}{the proposal rate parameter between 0 and 1}
}
\value{
the efficiency of the IMH
}
\description{
Efficiency (integrated autorrelation time) of the IMH with exponential target
}
