% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_LAR1.r
\name{LAR_autocorr}
\alias{LAR_autocorr}
\title{Computes the autocorrelation function of a Latent autoregressive (LAR1) for quantiles of a LNAR1 process}
\usage{
LAR_autocorr(l, a, phi, sigma_e)
}
\arguments{
\item{l}{lag}

\item{a}{the (log) quantile of interest}

\item{phi}{correlation parameter of the latent AR(1)}

\item{sigma_e}{the standard deviation of the innovations of the latent AR(1)}
}
\value{
the autocorrelation at lag \code{l}
}
\description{
Computes the autocorrelation function of a Latent autoregressive (LAR1) for quantiles of a LNAR1 process
}
\details{
the LAR is a \strong{binary} process.
This is mostly an internal function.
}
