% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_LNAR1.r
\name{LAR_autocorr_2}
\alias{LAR_autocorr_2}
\title{Computes the autocorrelation function of a Latent autoregressive (LAR1) for quantiles of a LNAR1 process}
\usage{
LAR_autocorr_2(l, a, phi, sigma_e)
}
\arguments{
\item{l}{lag}

\item{a}{the (log) quantile of interest}

\item{phi}{correlation parameter of the latent AR(1)}

\item{sigma_e}{the standard deviation of the innovations of the latent AR(1)}
}
\value{
the autocorrelation at lag \code{l}
}
\description{
Computes the autocorrelation function of a Latent autoregressive (LAR1) for quantiles of a LNAR1 process
}
\details{
the LAR is a \strong{binary} process.
This is mostly an internal function.
Here we are computing the autocorrelation of a process of the type
exp(Z_t) <= log(a), where Z_t is an AR(1, lmeam, lsd, phi).
}
