% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_LAR1.r
\name{LAR_ess}
\alias{LAR_ess}
\title{Compute the true effective sample size (ESS) for an LAR(1) process}
\usage{
LAR_ess(N, p, phi, sigma_e, K = 1000)
}
\arguments{
\item{N}{the number of draws}

\item{p}{the marginal probability that X = 1}

\item{phi}{correlation parameter of the latent AR(1)}

\item{sigma_e}{the standard deviation of the innovations of the latent AR(1)}

\item{K}{an upper bound up to which to sum the autocorrelation function.
Default is K = 5000}
}
\value{
a scalar with the ESS
}
\description{
Compute the true effective sample size (ESS) for an LAR(1) process
}
