% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_LNAR1.r
\name{LNAR1_eff}
\alias{LNAR1_eff}
\title{Compute the efficiency (for the mean) of a LNAR1 process}
\usage{
LNAR1_eff(lmean, lsd, phi, K = 1e+05)
}
\arguments{
\item{lmean}{log mean}

\item{lsd}{log standard deviation}

\item{phi}{correlation parameter of the latent AR(1)}

\item{K}{an upper bound up to which to sum the autocorrelation function.}
}
\value{
the efficiency, which should be >= 0
}
\description{
Compute the efficiency (for the mean) of a LNAR1 process
}
