% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_LNAR1.r
\name{LNAR_quantile_autocorr}
\alias{LNAR_quantile_autocorr}
\title{Autocorrelation function for the p-th quantile of a LNAR1 process}
\usage{
LNAR_quantile_autocorr(k, xi_p, lmean, lsd, phi, K)
}
\arguments{
\item{k}{lag}

\item{xi_p}{the \emph{value} of the p-th quantile}

\item{lmean}{log mean}

\item{lsd}{log standard deviation}

\item{phi}{correlation parameter of the latent AR(1)}

\item{K}{an upper bound up to which to sum the autocorrelation function.
Default is K = 5000}
}
\value{
the autocorrelation value
}
\description{
Autocorrelation function for the p-th quantile of a LNAR1 process
}
\details{
Here we are using the autocorrelation of a binary process (LAR1) to compute
the autocorrelation of a process of the type
exp(Z_t) <= log(a), where Z_t is an AR(1, lmean, lsd, phi).
}
