% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.r
\name{lognormal_hpd}
\alias{lognormal_hpd}
\title{Compute the highest density interval for a log-normal}
\usage{
lognormal_hpd(alpha = 0.95, lmean = 0, lsd = 1)
}
\arguments{
\item{alpha}{a number between 0 and 1 corresponding to the level of the BCI}

\item{lmean}{log-normal mean parameter (mu)}

\item{lsd}{log-normal standard deviation parameter (sigma)}
}
\value{
a vector of size 2 with the HDI/HPD
}
\description{
Compute the highest density interval for a log-normal
}
