% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.r
\name{lognormal_hpd_percentiles}
\alias{lognormal_hpd_percentiles}
\title{Compute the quantiles corresponding the the HPD endpoints in a log-normal}
\usage{
lognormal_hpd_percentiles(alpha = 0.95, lmean = 0, lsd = 1)
}
\arguments{
\item{alpha}{a number between 0 and 1 corresponding to the level of the BCI}

\item{lmean}{log-normal mean parameter (mu)}

\item{lsd}{log-normal standard deviation parameter (sigma)}
}
\value{
a vector of size 2 with the quantiles
}
\description{
Compute the quantiles corresponding the the HPD endpoints in a log-normal
}
