% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.r
\name{lognormal_mean}
\alias{lognormal_mean}
\title{Mean (expectation) of a log-normal}
\usage{
lognormal_mean(lmean = 0, lsd = 1)
}
\arguments{
\item{lmean}{log-normal mean parameter (mu)}

\item{lsd}{log-normal standard deviation parameter (sigma)}
}
\value{
the mean
}
\description{
Mean (expectation) of a log-normal
}
