% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.r
\name{lognormal_median}
\alias{lognormal_median}
\title{Median of a log-normal distribution}
\usage{
lognormal_median(lmean = 0)
}
\arguments{
\item{lmean}{log-normal mean parameter (mu)}
}
\value{
a scalar with the median
}
\description{
Median of a log-normal distribution
}
