% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_utilities.r
\name{lognormal_mode}
\alias{lognormal_mode}
\title{Mode of a log-normal}
\usage{
lognormal_mode(lmean = 0, lsd = 1)
}
\arguments{
\item{lmean}{log-normal mean parameter (mu)}

\item{lsd}{log-normal standard deviation parameter (sigma)}
}
\value{
the mode
}
\description{
Mode of a log-normal
}
