% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.r
\name{lognormal_skewness}
\alias{lognormal_skewness}
\title{Skewness of a log-normal distribution}
\usage{
lognormal_skewness(lsd = 1)
}
\arguments{
\item{lsd}{log-normal standard deviation parameter (sigma)}
}
\value{
a scalar with the skewness
}
\description{
Skewness of a log-normal distribution
}
