% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_utilities.r
\name{lognormal_variance}
\alias{lognormal_variance}
\title{Compute the variance of a log-normal distribution}
\usage{
lognormal_variance(lmean, lsd)
}
\arguments{
\item{lmean}{log mean}

\item{lsd}{log standard deviation}
}
\value{
the variance
}
\description{
Compute the variance of a log-normal distribution
}
\examples{
lognormal_variance(0, 1)
var(exp(rnorm(1e6)))
}
