% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_MH.r
\name{MH_exponential}
\alias{MH_exponential}
\title{Independence Metropolis-Hastings with exponential(1) target}
\usage{
MH_exponential(M = 10000, theta = 1, x0 = 0.1)
}
\arguments{
\item{M}{number of steps}

\item{theta}{proposal rate parameter}

\item{x0}{initial state}
}
\value{
a data.frame containing the states, the log target density and whether the move was accepted
}
\description{
Independence Metropolis-Hastings with exponential(1) target
}
