% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_LAR1.r
\name{generate_LAR_TS}
\alias{generate_LAR_TS}
\title{Sample from a LAR(1) process}
\usage{
generate_LAR_TS(N, phi, target.p)

generate_LAR_TS(N, phi, target.p)
}
\arguments{
\item{N}{number of draws}

\item{phi}{correlation parameter between -1 and 1}

\item{target.p}{target marginal probability that X = 1}
}
\value{
a vector with \code{N} draws

a vector of \code{N} binary draws
}
\description{
Sample from a LAR(1) process

Generate draws from a latent autorregressive (LAR1) model
}
\examples{
p <- .23
thePhi <- get_LAR_phi(Eff = .6, p = p)
X <- generate_LAR_TS(N = 1E3, phi = thePhi, target.p = p)
the.phi <- get_LAR_phi(Eff = .2, p = .12)
X <- generate_LAR_TS(N = 1e3, phi = the.phi, target.p = .12)
acf(X)
}
