% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_utilities.r
\name{lognormal_bci}
\alias{lognormal_bci}
\title{Compute the central BCI for a log-normal distribution}
\usage{
lognormal_bci(alpha, lmean, lsd)
}
\arguments{
\item{alpha}{the HPD level  (default is 0.95)}

\item{lmean}{log mean}

\item{lsd}{log standard deviation}
}
\value{
a vector of size two with the BCI endpoints
}
\description{
Compute the central BCI for a log-normal distribution
}
\examples{
lognormal_bci(0.95, 0, 1)
lognormal_bci(0.90, 0, 1)
}
