% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_utilities.r
\name{lognormal_hpd}
\alias{lognormal_hpd}
\title{Compute the HPD of a log-normal distribution}
\usage{
lognormal_hpd(alpha = 0.95, lmean, lsd)
}
\arguments{
\item{alpha}{the HPD level  (default is 0.95)}

\item{lmean}{log mean}

\item{lsd}{log standard deviation}
}
\value{
a vector of size two with the HPD endpoints
}
\description{
Compute the HPD of a log-normal distribution
}
\examples{
lognormal_hpd(0.95, 0, 1)
lognormal_hpd(0.90, 0, 1)
}
