% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_utilities.r
\name{lognormal_hpd_percentiles}
\alias{lognormal_hpd_percentiles}
\title{Get the percentiles corresponding to the HPD of a log-normal}
\usage{
lognormal_hpd_percentiles(alpha = 0.95, lmean, lsd)
}
\arguments{
\item{alpha}{the HPD level  (default is 0.95)}

\item{lmean}{log mean}

\item{lsd}{log standard deviation}
}
\value{
a vector with two probabilities
}
\description{
Get the percentiles corresponding to the HPD of a log-normal
}
\examples{
lognormal_hpd_percentiles(0.95, 0, 1)
lognormal_hpd_percentiles(0.90, 0, 1)
lognormal_hpd_percentiles(0.95, 0, .1)
lognormal_hpd_percentiles(0.90, 0, .1)
}
