% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_utilities.r
\name{lognormal_mean}
\alias{lognormal_mean}
\title{Compute the mean of a log-normal distribution}
\usage{
lognormal_mean(lmean, lsd)
}
\arguments{
\item{lmean}{log mean}

\item{lsd}{log standard deviation}
}
\value{
the mean
}
\description{
Compute the mean of a log-normal distribution
}
\examples{
lognormal_mean(0, 1)
mean(exp(rnorm(1e6)))
}
