% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_utilities.r
\name{lognormal_median}
\alias{lognormal_median}
\title{Compute the median of a log-normal distribution}
\usage{
lognormal_median(lmean, lsd)
}
\arguments{
\item{lmean}{log mean}

\item{lsd}{log standard deviation}
}
\value{
the median
}
\description{
Compute the median of a log-normal distribution
}
\examples{
lognormal_median(0, 1)
median(exp(rnorm(1e6)))
}
