% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_utilities.r
\name{lognormal_skewness}
\alias{lognormal_skewness}
\title{Compute the skewness of a log-normal}
\usage{
lognormal_skewness(lsd)
}
\arguments{
\item{lsd}{log standard deviation}
}
\value{
a scalar with the skewness
}
\description{
Compute the skewness of a log-normal
}
\examples{
lognormal_skewness(.1)
lognormal_skewness(.5)
lognormal_skewness(1)
curve(lognormal_skewness)
}
